# application/services/custom_services/tickets/src/models/entitymodels/tickets.ts
export interface tickets 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   description: { type: String },
   types: { type: ObjectId, ref: 'types' }
}